solve add a building1

+++ controllers/Parking/deleteParkingSlot.js
@@ -1,8 +1,18 @@
-import { Parkings } from "../../models";
+import { Building, Parkings } from "../../models";
 import { catchAsyncError } from "../../utility";
 import errorHandler from "../../utility/errorHandlerClass";
 
 export const deleteParkingSlot = catchAsyncError(async (req, res, next) => {
+  const managerEmail = req.user.email;
+
+  const building = await Building.findOne({ managerEmail: managerEmail });
+
+  if (!building) {
+    return res.status(400).json({
+      message: "You are not authorised!",
+    });
+  }
+
   const { id } = req.params;
 
   const parkingSlot = await Parkings.findByIdAndDelete({ _id: id });
@@ -12,6 +22,10 @@ export const deleteParkingSlot = catchAsyncError(async (req, res, next) => {
       new errorHandler(`A parking slot with ID: ${id}, not found`, 404)
     );
   }
+
+  building.availableSpots = parseInt(building.availableSpots) - 1;
+
+  await building.save();
 
   res.status(200).json({
     message: `A parking slot with ID: ${id}, deleted successfully!`,
